auth middleware created | some file edits
