+++ routes/auth-route.js
@@ -1,7 +1,6 @@
+const authController = require("../controller/auth-controller");
 const express = require("express");
 const router = express.Router();
-
-const authController = require("../controller/auth-controller");
 
 router.post("/sign-up", authController.signUp);
 router.post("/log-in", authController.login);
