+++ routes/commet-route.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
 
-const auth = require("../controller/auth-controller");
+const check = require("../middleware/auth");
 const commentController = require("../controller/comments-controller");
 
-router.post("/:postId", auth.isAuth, commentController.addComment);
+router.post("/:postId", check.isAuthenticated, commentController.addComment);
 
 module.exports = router;
