+++ routes/post-route.js
@@ -1,12 +1,11 @@
 const express = require("express");
 const router = express.Router();
 
-const authController = require("../controller/auth-controller");
+const check = require("../middleware/auth");
 const postsController = require("../controller/posts-controller");
 
-router.get("/page/:page", postsController.getPost);
-router.post("/", authController.isAuth, postsController.addPost);
-router.get("/:id", postsController.getSinglePost);
+router.post("/", check.isAuthenticated, postsController.addPost);
+router.get("/:id", check.isAuthenticated, postsController.getSinglePost);
 router.get("/user/:id", postsController.getUserPost);
 
 module.exports = router;
